Redo all of the simulation.py, the old one moved to as a simulation (previous).py for a backup and reference. Added a new sound file including BGM, and effects Create an assets for images and logo
