+++ balancescale/initialize/display.py
@@ -6,7 +6,7 @@ screen = pygame.display.set_mode((width, height))
 pygame.display.set_caption('Balance Scale')
 
 #Set the icon
-icon = pygame.image.load("balancescale/assets/images/balancescale.png")
+icon = pygame.image.load("balancescale/assets/images/Logo.png")
 pygame.display.set_icon(icon)
 
 #Set up background
